Refactor color controls to use GeneratePressControl
