+++ src/customizer-controls/color-manager/index.js
@@ -1,4 +1,7 @@
-import GeneratePressColorManagerControl from './GeneratePressColorManagerControl';
+import GeneratePressControl from '../../components/GeneratePressControl';
+import GeneratePressColorManagerControlForm from './GeneratePressColorManagerControlForm';
+
+const GeneratePressColorManagerControl = GeneratePressControl.extend( GeneratePressColorManagerControlForm );
 
 // Register control type with Customizer.
 wp.customize.controlConstructor[ 'generate-color-manager-control' ] = GeneratePressColorManagerControl;
